Use nbconvert for the functional tests
