--- functional_tests/functional_tests_examples/README.txt
@@ -1,7 +1,0 @@
-This test takes "bluepysnap/doc/source/notebooks" examples and run
-them programmatically using IPython lib. Results from the tests are
-compared to the original notebook results.
-It only compares png and text/html outputs.
-You must have these extra python libraries installed :
-- jupyter
-- imageio
